refactor(frontend): split root layout chrome
