+++ apps/frontend/src/routes/+layout.svelte
@@ -1,30 +1,16 @@
 <script lang="ts">
-  import { afterNavigate, goto } from '$app/navigation';
-  import { resolve } from '$app/paths';
   import { page } from '$app/state';
-  import * as m from '$lib/i18n/messages';
-  import { onNotificationClick } from '$lib/notifications/pushNotifications';
-  import { prepareUiForNotificationPath } from '$lib/notifications/notificationNavigationUi';
-  import ServerGutter from '$lib/ServerGutter.svelte';
   import { setAuthServerInfo } from '$lib/components/authServerInfo';
-  import ConnectionIndicator from '$lib/components/ConnectionIndicator.svelte';
   import ConnectionProvider from '$lib/components/ConnectionProvider.svelte';
   import GlobalKeyboardShortcuts from '$lib/components/GlobalKeyboardShortcuts.svelte';
   import IdleTracker from '$lib/components/IdleTracker.svelte';
   import UpdateNotifier from '$lib/components/UpdateNotifier.svelte';
-  import { usePageTitle, usePinchZoomPrevention, useVisualViewport } from '$lib/hooks';
-  import {
-    SIDEBAR_PANEL_WIDTH_PX,
-    sidebarEdgeSwipe,
-    sidebarSwipe
-  } from '$lib/hooks/useSidebarSwipe.svelte';
-  import { chatRoomIdFromRoute } from '$lib/navigation/chatRoomRoute';
-  import { getActiveServer } from '$lib/state/activeServer.svelte';
-  import { sidebarNav } from '$lib/state/globals.svelte';
+  import { usePageTitle } from '$lib/hooks';
   import { provideAppUiState } from '$lib/state/appUi.svelte';
   import { useServerRegistry } from '$lib/state/server/useServerRegistry.svelte';
   import { ToastContainer } from '$lib/ui/toast';
-  import { AppHeader, Frame } from '$lib/ui';
+  import RootLayoutEffects from './RootLayoutEffects.svelte';
+  import RootLayoutFrame from './RootLayoutFrame.svelte';
   import '../app.css';
 
   let { data, children } = $props();
@@ -38,49 +24,12 @@
   setAuthServerInfo(() => data.serverInfo);
   const appUi = provideAppUiState();
   useServerRegistry(() => data.user);
-  useVisualViewport();
-  usePinchZoomPrevention();
-
-  const activeServerId = $derived(getActiveServer());
-  const activeRoomId = $derived(chatRoomIdFromRoute(page.route.id, page.params.roomId));
-
-  $effect(() => {
-    if (typeof activeRoomId === 'string' && activeRoomId) {
-      appUi.setActiveRoomScope(activeServerId, activeRoomId);
-      return;
-    }
-    appUi.setActiveServer(activeServerId);
-  });
-
-  // Route push-notification clicks via SvelteKit's client-side navigation
-  // instead of letting the SW do a full document navigation. Same-URL
-  // clicks become a no-op; cross-URL clicks just update the route.
-  $effect(() =>
-    onNotificationClick((url) => {
-      try {
-        const target = new URL(url);
-        if (target.origin !== window.location.origin) return;
-        prepareUiForNotificationPath(appUi, target.pathname);
-        return goto(resolve((target.pathname + target.search + target.hash) as '/'));
-      } catch {
-        // Ignore malformed URLs from the SW.
-      }
-    })
-  );
-
-  $effect(() => sidebarNav.initViewportTracking());
-  afterNavigate(() => {
-    if (sidebarNav.isMobile) sidebarNav.close();
-  });
 
   const getFullTitle = usePageTitle();
   const fullTitle = $derived(getFullTitle());
-
-  function stopPropagation(event: Event) {
-    event.stopPropagation();
-  }
 </script>
 
+<RootLayoutEffects {appUi} />
 <GlobalKeyboardShortcuts />
 <IdleTracker />
 <UpdateNotifier />
@@ -90,90 +39,10 @@
 </svelte:head>
 
 <ConnectionProvider>
-  {@render frame()}
+  <RootLayoutFrame>
+    {@render children?.()}
+  </RootLayoutFrame>
 </ConnectionProvider>
-
-{#snippet frame()}
-  {@const progress = sidebarNav.isMobile ? sidebarNav.progress : 1}
-  {@const dragging = sidebarNav.dragOffset !== null}
-  {@const mobileClosed = sidebarNav.isMobile && progress === 0 && !dragging}
-  {@const tx = (progress - 1) * SIDEBAR_PANEL_WIDTH_PX}
-  <div
-    class="flex h-full w-full flex-col overscroll-y-contain bg-surface-100 pt-[env(safe-area-inset-top,0px)] md:p-3 md:pt-0"
-  >
-    <ConnectionIndicator />
-
-    <AppHeader />
-
-    <Frame class="relative flex-col">
-      {#if sidebarNav.isMobile}
-        <!--
-          Edge gesture zone (swipe-to-open). `touch-action: none` is essential:
-          without it, Chrome / iOS Safari fire pointercancel ~8px into a
-          horizontal drag (text-selection / back-navigation gesture detection).
-          Hidden when sidebar is open (the backdrop takes over). Plain taps are
-          intentionally swallowed here; this target exists only to start swipes.
-        -->
-        {#if !sidebarNav.isOpen || dragging}
-          <div
-            use:sidebarEdgeSwipe
-            data-app-sidebar="true"
-            data-testid="mobile-sidebar-edge"
-            class="fixed top-11 bottom-0 left-0 z-40 w-6 touch-none md:hidden"
-            aria-hidden="true"
-            onpointerdown={stopPropagation}
-            onpointerup={stopPropagation}
-            onclick={stopPropagation}
-            oncontextmenu={stopPropagation}
-          ></div>
-        {/if}
-
-        <button
-          type="button"
-          use:sidebarSwipe
-          data-app-sidebar="true"
-          data-testid="mobile-sidebar-backdrop"
-          class={[
-            'fixed inset-0 top-11 z-40 touch-none bg-black/50 md:hidden',
-            !dragging && 'transition-opacity duration-200',
-            mobileClosed && 'pointer-events-none'
-          ]}
-          style:opacity={progress}
-          disabled={mobileClosed}
-          tabindex={mobileClosed ? -1 : 0}
-          aria-hidden={mobileClosed}
-          onclick={() => sidebarNav.close()}
-          aria-label={m['common.close_sidebar']()}
-        ></button>
-      {/if}
-
-      <div class="flex min-h-0 flex-1 flex-row">
-        <div
-          use:sidebarSwipe
-          data-app-sidebar="true"
-          data-testid="mobile-sidebar-panel"
-          class={[
-            'z-50 min-h-0 flex-col self-stretch bg-background',
-            'max-md:fixed max-md:top-11 max-md:bottom-0 max-md:left-0 max-md:w-17 max-md:touch-pan-y',
-            // Mobile: always rendered so we can animate transform.
-            // Desktop: hide entirely when closed (no overlay; layout reflows).
-            sidebarNav.isMobile ? 'flex' : sidebarNav.isOpen ? 'flex' : 'hidden',
-            // Mobile-only: hide via `visibility: hidden` after the close
-            // transition, so Playwright / accessibility tooling correctly see
-            // the sidebar as not-visible while the slide-out animation works.
-            mobileClosed && 'sidebar-mobile-closed',
-            !dragging && 'sidebar-mobile-anim'
-          ]}
-          style:transform={sidebarNav.isMobile ? `translateX(${tx}px)` : undefined}
-        >
-          <ServerGutter />
-        </div>
-
-        {@render children?.()}
-      </div>
-    </Frame>
-  </div>
-{/snippet}
 
 {#if page.state.modal}
   {#await loadModalContainer() then { default: ModalContainer }}
@@ -182,29 +51,3 @@
 {/if}
 
 <ToastContainer />
-
-<style>
-  /*
-    Mobile sidebar animation — slide via transform, plus a delayed visibility
-    swap so the off-screen panel is reported as `visibility: hidden` (not just
-    visually hidden by transform) once the close animation finishes. This
-    matters for accessibility tooling and Playwright's `toBeVisible()`.
-
-    Open  → transform animates 200ms, visibility flips to `visible` immediately.
-    Close → transform animates 200ms, visibility flips to `hidden` AFTER 200ms.
-  */
-  @media (max-width: 767px) {
-    :global(.sidebar-mobile-anim) {
-      visibility: visible;
-      transition:
-        transform 200ms ease-out,
-        visibility 0s linear 0s;
-    }
-    :global(.sidebar-mobile-anim.sidebar-mobile-closed) {
-      visibility: hidden;
-      transition:
-        transform 200ms ease-out,
-        visibility 0s linear 200ms;
-    }
-  }
-</style>
